protect controllers with policy

+++ app/Http/Controllers/API/Chat/DestroyController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API\Chat;
 use App\Http\Controllers\Controller;
 use App\Models\Chat;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 use Throwable;
 
 class DestroyController extends Controller
@@ -14,6 +15,8 @@ class DestroyController extends Controller
      */
     public function __invoke(Chat $chat): JsonResponse
     {
+        Gate::authorize('forceDelete', $chat);
+
         $chat->deleteOrFail();
 
         return response()->json(['message' => 'OK']);
